Support old syntax:
This commit adds support for the previous keywords, seamlessly
implementing them in the new architecture.

"register" only needs to register within the right variable, so
it does.

"register_plugin" is superfluous so it just doesn't do anything.

"on_plugin_import" registers additional callbacks to be called
later. It registers them to be called with the plugin variable
as the first parameter, which is what methods are actually for.

In the new architecture you simply have methods. Here we must save
them for later, and then call them again. We have to save them in
a variable that can be reached, so we create a method to reach a
lexical variables instead of a global. The variable name attempt
to be as unique as possible.

Perhaps need to rename it, since all caps is the convention for
globals, but it works.

No warnings for deprecation are added yet. This seems to work but
additional tests are necessary.